Add very basic support for AY music files.
ZXAYEMUL files contain Z80 code for the generation of music and sound
effects.  They can be created by copying the sound generating parts of
programs to a new file.  A Z80 emulator is required to play ZXAYEMUL
files.  A ZXAYEMUL file contains blocks of code that are copied into the
Z80's address space and executed.  This commit will load ZXAYEMUL files
to the Z80's address space, set the Z80 registers and attempt to play
the file by emulating the Z80.  It works for a small fraction of
available ZXAYEMUL files at present.